+++ packages/core/src/interfaces/index.ts
@@ -1,2 +1,3 @@
 export * from './BasePluginService.js'
 export * from './BaseEntity.js'
+export * from './BaseSchemaEntity.js'