+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @ayazmo/utils
 
+## 0.5.0
+
+### Minor Changes
+
+- switch to entityschema
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ayazmo/types@0.5.0
+
 ## 0.4.1
 
 ### Patch Changes